fix: refactor home screen code
