+++ frontend/app/src/main/java/com/example/haengsha/MainActivity.kt
@@ -1,10 +1,8 @@
 package com.example.haengsha
 
-import android.os.Build
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.annotation.RequiresApi
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
