+++ frontend/app/src/main/java/com/example/haengsha/model/network/dataModel/SignupInfo.kt
@@ -1,4 +1,4 @@
-package com.example.haengsha.model.dataSource
+package com.example.haengsha.model.network.dataModel
 
 object SignupInfo {
     val college = listOf(
