+++ frontend/app/src/main/java/com/example/haengsha/ui/screens/home/Home.kt
@@ -33,14 +33,12 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.example.haengsha.R
+import com.example.haengsha.model.network.dataModel.EventResponse
 import com.example.haengsha.model.uiState.UserUiState
-import com.example.haengsha.model.viewModel.event.EventApiViewModel
-import com.example.haengsha.model.viewModel.event.RecommendationApiViewModel
+import com.example.haengsha.model.viewModel.home.HomeApiViewModel
+import com.example.haengsha.model.viewModel.home.HomeViewModel
 import com.example.haengsha.ui.theme.HaengshaBlue
 import com.example.haengsha.ui.theme.poppins
 import com.kizitonwose.calendar.compose.WeekCalendar
@@ -56,56 +54,42 @@ import java.time.format.TextStyle
 import java.util.Date
 import java.util.Locale
 
+private val dateFormatter = DateTimeFormatter.ofPattern("dd")
+private val monthFormatter = DateTimeFormatter.ofPattern("MMMM")
 
-class SharedViewModel : ViewModel() {
-    private val _selectedDate = MutableLiveData(LocalDate.now())
-    val selectedDate: LiveData<LocalDate> = _selectedDate
-    private val _festivalItems = MutableLiveData<List<EventCardData>?>()
-    private val _academicItems = MutableLiveData<List<EventCardData>?>()
-    val festivalItems: MutableLiveData<List<EventCardData>?> = _festivalItems
-    val academicItems: MutableLiveData<List<EventCardData>?> = _academicItems
-
-    fun updateSelectedDate(newDate: LocalDate) {
-        _selectedDate.value = newDate     // Update functions to set LiveData properties
-    }
-
-    fun updateEventItems(festivalItems: List<EventCardData>?, academicItems: List<EventCardData>?) {
-        _festivalItems.value = festivalItems
-        _academicItems.value = academicItems
-    }
+@Composable
+fun Home(
+    homeViewModel: HomeViewModel,
+    homeApiViewModel: HomeApiViewModel,
+    innerPadding: PaddingValues,
+    userUiState: UserUiState
+) {
+    HomeScreen(homeApiViewModel, innerPadding, userUiState, homeViewModel)
 }
-
 
 @SuppressLint("UnusedMaterialScaffoldPaddingParameter")
 @Composable
 fun HomeScreen(
-    recommendationApiViewModel: RecommendationApiViewModel,
+    homeApiViewModel: HomeApiViewModel,
     innerPadding: PaddingValues,
     userUiState: UserUiState,
-    sharedViewModel: SharedViewModel
+    homeViewModel: HomeViewModel
 ) {
-    val eventApiViewModel: EventApiViewModel =
-        viewModel(factory = EventApiViewModel.Factory(sharedViewModel))
     var selection by remember { mutableStateOf(LocalDate.now()) }
     val currentMonth = selection.yearMonth
     val startDate = remember { currentMonth.minusMonths(100).atStartOfMonth() } // Adjust as needed
     val endDate = remember { currentMonth.plusMonths(100).atEndOfMonth() } // Adjust as needed
     val firstDayOfWeek = remember { firstDayOfWeekFromLocale() } // Available from the library
 
-    recommendationApiViewModel.getRecommendationList(token = userUiState.token)
-
-    eventApiViewModel.getEventByDate(selection)
+    homeApiViewModel.getRecommendationList(token = userUiState.token)
+    
+    homeApiViewModel.getEventByDate(selection)
     val state = rememberWeekCalendarState(
         startDate = startDate,
         endDate = endDate,
         firstVisibleWeekDate = selection,
         firstDayOfWeek = firstDayOfWeek,
     )
-
-
-    var date by remember {
-        mutableStateOf("Open date picker dialog")
-    }
 
     var showDatePicker by remember {
         mutableStateOf(false)
@@ -176,35 +160,13 @@ fun HomeScreen(
                 Day(day.date, isSelected = selection == day.date) { clicked ->
                     if (selection != clicked) {
                         selection = clicked
-                        eventApiViewModel.getEventByDate(selection)
+                        homeApiViewModel.getEventByDate(selection)
                     }
                 }
             },
         )
 
-        TabView(sharedViewModel, recommendationApiViewModel, selection)
-
-
-    }
-}
-
-
-private val dateFormatter = DateTimeFormatter.ofPattern("dd")
-
-private val monthFormatter = DateTimeFormatter.ofPattern("MMMM")
-
-fun parseStringToDate(dateString: String): LocalDate? {
-    val formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd")
-    return try {
-        LocalDate.parse(dateString, formatter)
-    } catch (e: Exception) {
-        null // Handle the case when the input string is not in the expected format
-    }
-}
-
-fun DayOfWeek.displayText(uppercase: Boolean = false): String {
-    return getDisplayName(TextStyle.SHORT, Locale.ENGLISH).let { value ->
-        if (uppercase) value.uppercase(Locale.ENGLISH) else value
+        TabView(homeViewModel, homeApiViewModel)
     }
 }
 
@@ -261,9 +223,7 @@ fun MyDatePickerDialog(
         onDismissRequest = { onDismiss() },
         confirmButton = {
             Button(onClick = {
-                if (selectedDate != null) {
-                    onDateSelected(selectedDate)
-                }
+                onDateSelected(selectedDate)
                 onDismiss()
             }
 
@@ -285,30 +245,72 @@ fun MyDatePickerDialog(
     }
 }
 
+fun parseStringToDate(dateString: String): LocalDate? {
+    val formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd")
+    return try {
+        LocalDate.parse(dateString, formatter)
+    } catch (e: Exception) {
+        null // Handle the case when the input string is not in the expected format
+    }
+}
+
+fun DayOfWeek.displayText(uppercase: Boolean = false): String {
+    return getDisplayName(TextStyle.SHORT, Locale.ENGLISH).let { value ->
+        if (uppercase) value.uppercase(Locale.ENGLISH) else value
+    }
+}
+
 private fun convertMillisToDate(millis: Long): String {
-    val formatter = SimpleDateFormat("yyyy-MM-dd")
+    val formatter = SimpleDateFormat("yyyy-MM-dd", Locale.ROOT)
     return formatter.format(Date(millis))
 }
 
-@Composable
-fun Home(
-    recommendationApiViewModel: RecommendationApiViewModel,
-    innerPadding: PaddingValues,
-    userUiState: UserUiState
-) {
-    val sharedViewModel = viewModel<SharedViewModel>()
+fun EventResponse.toEventCardData(): EventCardData {
+    val startDate = stringToDate(eventDurations[0].eventDay)
+    var endDate = startDate
+    if (eventDurations.size > 1) {
+        endDate = stringToDate(eventDurations[eventDurations.size - 1].eventDay)
+    }
 
-    HomeScreen(recommendationApiViewModel, innerPadding, userUiState, sharedViewModel)
+    var eventType = "Festival"
+    if (!isFestival) {
+        eventType = "Academic"
+    }
+    return EventCardData(
+        organizer = author.nickname,
+        eventTitle = title,
+        startDate = startDate,
+        endDate = endDate,
+        likes = likeCount,
+        favorites = favoriteCount,
+        eventType = eventType,
+        place = place,
+        time = time ?: "wow"
+    )
+}
+
+fun stringToDate(dateString: String): LocalDate {
+    val formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd")
+    try {
+        return LocalDate.parse(dateString, formatter)
+        //val format = SimpleDateFormat("yyyy-MM-dd HH:mm:ss")
+        //date = format.parse(dateString)!!
+    } catch (e: Exception) {
+        e.printStackTrace()
+    }
+    return LocalDate.now()
 }
 
 @Preview(showBackground = true)
 @Composable
 fun HomeScreenPreview() {
-    val recommendationApiViewModel: RecommendationApiViewModel =
-        viewModel(factory = RecommendationApiViewModel.Factory)
+    val homeViewModel = viewModel<HomeViewModel>()
+    val homeApiViewModel: HomeApiViewModel =
+        viewModel(factory = HomeApiViewModel.Factory(homeViewModel))
 
     Home(
-        recommendationApiViewModel = recommendationApiViewModel,
+        homeViewModel = homeViewModel,
+        homeApiViewModel = homeApiViewModel,
         innerPadding = PaddingValues(0.dp),
         userUiState = UserUiState("foo", "bar")
     )
