+++ frontend/app/src/main/java/com/example/haengsha/ui/screens/login/LoginScreen.kt
@@ -130,6 +130,9 @@ fun LoginScreen(
         Spacer(modifier = Modifier.height(50.dp))
         CommonBlueButton(text = "로그인하기",
             onClick = {
+                emailInput = "user2@snu.ac.kr"
+                passwordInput = "user2"
+
                 if (emailInput.trimStart() == "") {
                     isEmailError = true
                     Toasty.error(
